+++ src/main/java/nhbank/core/domain/ACOM_COM_FPBODYInfo.java
@@ -7,7 +7,8 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.Table;
 import java.io.Serializable;
-import java.sql.Date;
+import java.math.BigDecimal;
+import java.util.Date;
 
 @Entity
 @Table(name = "acomComFpbody")
@@ -16,9 +17,9 @@ public class ACOM_COM_FPBODYInfo implements Serializable {
     @Id
     @GeneratedValue
     private String refNo;
-    private int hisNo;
+    private BigDecimal hisNo;
     private String prtKind;
-    private int lineSeq;
+    private BigDecimal lineSeq;
     private String lineTxt;
     private String regEmpNo;
     private Date regDt;
