+++ src/main/java/nhbank/core/domain/ACOM_COM_FPHEADInfo.java
@@ -7,7 +7,8 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.Table;
 import java.io.Serializable;
-import java.sql.Date;
+import java.math.BigDecimal;
+import java.util.Date;
 
 @Entity
 @Table(name = "acomComFphead")
@@ -16,7 +17,7 @@ public class ACOM_COM_FPHEADInfo implements Serializable {
     @Id
     @GeneratedValue
     private String refNo;
-    private int hisNo;
+    private BigDecimal hisNo;
     private String prtKind;
     private String upmuCd;
     private String trCd;
