+++ src/main/java/nhbank/core/domain/ACOM_LMT_SILHISInfo.java
@@ -7,7 +7,8 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.Table;
 import java.io.Serializable;
-import java.sql.Date;
+import java.math.BigDecimal;
+import java.util.Date;
 
 @Entity
 @Table(name = "acomLmtSilhis")
@@ -18,7 +19,7 @@ public class ACOM_LMT_SILHISInfo implements Serializable {
     private String sngNo;
     private String actCd;
     private String silCcy;
-    private int hisNo;
+    private BigDecimal hisNo;
     private String sts;
     private String trCd;
     private Date ibIl;
@@ -28,26 +29,26 @@ public class ACOM_LMT_SILHISInfo implements Serializable {
     private String silGb;
     private String upmuCd;
     private String georCd;
-    private int silAmt;
+    private BigDecimal silAmt;
     private String silPmgb;
-    private int silJan;
-    private int djgAmt;
+    private BigDecimal silJan;
+    private BigDecimal djgAmt;
     private String djgPmgb;
-    private int djgJan;
-    private int gurBamt;
+    private BigDecimal djgJan;
+    private BigDecimal gurBamt;
     private String gurBpmgb;
-    private int gurBjan;
-    private int gurFamt;
+    private BigDecimal gurBjan;
+    private BigDecimal gurFamt;
     private String gurFpmgb;
-    private int gurFjan;
+    private BigDecimal gurFjan;
     private String trBr;
     private String sdBr;
     private String trTime;
     private String opNo;
     private String mngNo;
-    private int trscSeqNo;
-    private int trxFamt;
-    private int trxBamt;
+    private BigDecimal trscSeqNo;
+    private BigDecimal trxFamt;
+    private BigDecimal trxBamt;
     private String regEmpNo;
     private Date regDt;
     private String regTm;
