+++ src/main/java/nhbank/core/domain/ADST_DPB_CERTInfo.java
@@ -7,7 +7,8 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.Table;
 import java.io.Serializable;
-import java.sql.Date;
+import java.math.BigDecimal;
+import java.util.Date;
 
 @Entity
 @Table(name = "adstDpbCert")
@@ -17,7 +18,7 @@ public class ADST_DPB_CERTInfo implements Serializable {
     @GeneratedValue
     private String acctNo;
     private Date ibIl;
-    private int txTrxNo;
+    private BigDecimal txTrxNo;
     private String sts;
     private String prtKind;
     private String drOpno;
