+++ src/main/java/nhbank/core/model/AFEX_XPH_SHIP_DTO.java
@@ -2,12 +2,13 @@ package nhbank.core.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
 public class AFEX_XPH_SHIP_DTO {
     private String REF_NO;
-    private int SEQ;
+    private BigDecimal SEQ;
     private String REG_EMP_NO;
     private Date REG_DT;
     private String REG_TM;
